move the package + combine java and kotlin sources
