+++ src/main/java/studio/dreamys/prometheus/essential/mixin/gg/essential/main/MixinBootstrap.java
@@ -1,11 +1,11 @@
-package studio.dreamys.prometheus.mixin.gg.essential.main;
+package studio.dreamys.prometheus.essential.mixin.gg.essential.main;
 
 import gg.essential.main.Bootstrap;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import studio.dreamys.prometheus.serial.EssentialCosmeticsManager;
+import studio.dreamys.prometheus.essential.serial.EssentialCosmeticsManager;
 
 @Mixin(value = Bootstrap.class, remap = false)
 public class MixinBootstrap {
